Don't start docker interactively when testing generated images
When stuffer tests the generated docker image, it executes docker
interactively, causing docker to expect to be connected to a tty.
This causes stuffer to fail if it is executed without a tty, for
example when calling it from cron.

This commit modifies the call to docker so that it does not expect
to be connected to a tty.