gantry: Announce new images and errors via pubsub messages
Gantry does not provide any feedback other than its return value,
leaving it up to the caller to integrate it with other applications.

This commit modifies gantry so that it announces images and errors
using pubsub messaging if the caller has specified topics for such
announcements on the command line.